Fixed issue due to swagger file improvement needed

+++ src/main/java/sibModel/RemainingCreditModelReseller.java
@@ -29,12 +29,12 @@ import java.io.IOException;
 @javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2020-03-23T17:54:09.105+05:30")
 public class RemainingCreditModelReseller {
   @SerializedName("sms")
-  private Long sms = null;
+  private Double sms = null;
 
   @SerializedName("email")
-  private Long email = null;
+  private Double email = null;
 
-  public RemainingCreditModelReseller sms(Long sms) {
+  public RemainingCreditModelReseller sms(Double sms) {
     this.sms = sms;
     return this;
   }
@@ -44,15 +44,15 @@ public class RemainingCreditModelReseller {
    * @return sms
   **/
   @ApiModelProperty(example = "12900", required = true, value = "SMS Credits remaining for reseller account")
-  public Long getSms() {
+  public Double getSms() {
     return sms;
   }
 
-  public void setSms(Long sms) {
+  public void setSms(Double sms) {
     this.sms = sms;
   }
 
-  public RemainingCreditModelReseller email(Long email) {
+  public RemainingCreditModelReseller email(Double email) {
     this.email = email;
     return this;
   }
@@ -62,11 +62,11 @@ public class RemainingCreditModelReseller {
    * @return email
   **/
   @ApiModelProperty(example = "2000000", required = true, value = "Email Credits remaining for reseller account")
-  public Long getEmail() {
+  public Double getEmail() {
     return email;
   }
 
-  public void setEmail(Long email) {
+  public void setEmail(Double email) {
     this.email = email;
   }
